refactor (ontologies): rename resource class file
