+++ src/models/v2/ontologies/class-definition.ts
